refactor(gateway): extract turn-end funnel to turn-end.ts (#2996 P8 PR-B)
Move the turn-END funnel + drain chokepoint (endCurrentTurnAtomic,
purgeReactionTracking, releaseTurnBufferGate, performBufferDrain,
drainBufferedIfAllowed, armNoReplyDrainTimer — the 4757-5271 family) VERBATIM
into telegram-plugin/gateway/turn-end.ts, parameterised by a TurnEndDeps
builder (gatewayTurnEndDeps(), exact-by-construction via ReturnType — the
stream-render idiom). State stays in gateway.ts; logic moves.

- M2 (design amendment): the purge tail's pending-restart drain crosses in the
  deps EXPLICITLY (pendingRestarts / triggerSelfRestart / turnInFlightForGate)
  so a queued /restart still fires at turn-complete — pinned by harness case 8.
- P9 getter funnel (clusters B/D): the module-scope currentTurn/inboundSpool/
  ipcServer reads in the moved bodies cross as live getters, leaving gateway.
- gateway.ts keeps thin function-declaration wrappers, so every callsite
  (stream-render deps, halt, disconnect-flush, silence-poke fallback,
  executeReply, the test seam) is byte-identical. Only 3 body rewrites:
  ipcServer.sendToAgent -> deps.sendToAgent, inboundSpool ->
  deps.getInboundSpool(), currentTurn?.replyCalled -> getCurrentTurn()?....
- Source-scrape suites retargeted in the same PR (the CLAUDE.md rule):
  buffer-gate-broadened, no-reply-bounded-drain, busy-ack-wiring,
  multitopic-routing-wiring, per-topic-current-turn, feed-heartbeat-liveness-open.
- gateway-line-ratchet LOWERED 26456 -> 26171 (gateway.ts now 26122 lines).

Refs #2996. Job: reference/jobs/on-the-leash.md.
