+++ scripts/gateway-line-ratchet.txt
@@ -1,1 +1,1 @@
-26456
+26171
